cleanup

+++ metagraph/src/common/hash/bloom_filter.hpp
@@ -9,11 +9,11 @@
 
 class BloomFilter {
   public:
-    BloomFilter(size_t filter_size = 0, size_t num_hash_functions = 0);
+    BloomFilter(size_t filter_size = 0, uint32_t num_hash_functions = 0);
 
     BloomFilter(size_t filter_size,
                 size_t expected_num_elements,
-                size_t max_num_hash_functions);
+                uint32_t max_num_hash_functions);
 
     void insert(uint64_t hash);
     bool check(uint64_t hash) const;
@@ -24,7 +24,7 @@ class BloomFilter {
     bool load(std::istream &in);
 
     size_t size() const { return filter_.size(); }
-    size_t num_hash_functions() const { return num_hash_functions_; }
+    uint32_t num_hash_functions() const { return num_hash_functions_; }
 
     const sdsl::bit_vector& data() const { return filter_; }
 
@@ -36,14 +36,14 @@ class BloomFilter {
         return -std::log2(false_positive_prob) * expected_num_elements / M_LN2;
     }
 
-    constexpr static size_t optim_h(double false_positive_prob) {
+    constexpr static uint32_t optim_h(double false_positive_prob) {
         if (false_positive_prob <= 0.0)
             throw std::runtime_error("False positive probability must be > 0.0");
 
         return std::ceil(-std::log2(false_positive_prob));
     }
 
-    constexpr static size_t optim_h(size_t filter_size, size_t expected_num_elements) {
+    constexpr static uint32_t optim_h(size_t filter_size, size_t expected_num_elements) {
         return expected_num_elements
             ? std::ceil(M_LN2 * filter_size / expected_num_elements)
             : static_cast<size_t>(-1);
@@ -51,7 +51,7 @@ class BloomFilter {
 
   private:
     sdsl::bit_vector filter_;
-    size_t num_hash_functions_;
+    uint32_t num_hash_functions_;
 };
